resolve tree test
