+++ client/src/components/App/__tests__/App.test.tsx
@@ -23,8 +23,8 @@ describe('The App Root Component', () => {
   });
 
   it('sets Test Suite state on mount', () => {
-    const mock = vi.spyOn(testSuitesApi, 'getTestSuites');
-    mock.mockResolvedValue(testSuites);
+    const getTestSuites = vi.spyOn(testSuitesApi, 'getTestSuites');
+    getTestSuites.mockResolvedValue(testSuites);
 
     render(
       <SnackbarProvider>
@@ -32,14 +32,7 @@ describe('The App Root Component', () => {
       </SnackbarProvider>
     );
 
-    // We have to wait for something to load so we don't get act()
-    // warnings.  The only thing rendered by App is children components.
-    // So await for those to be done with side effects (hooks).
-
-    // TODO: Render with new router
-    // await screen.findByText('Suite One');
-
-    expect(mock).toBeCalledTimes(1);
+    expect(getTestSuites).toBeCalledTimes(1);
   });
 
   it('sets the Test Session if there is a single Test Suite', () => {
@@ -48,18 +41,16 @@ describe('The App Root Component', () => {
 
     const postTestSessions = vi.spyOn(testSessionApi, 'postTestSessions');
     postTestSessions.mockResolvedValue(testSession);
+
     render(
       <SnackbarProvider>
         <App />
       </SnackbarProvider>
     );
 
-    // We have to wait for something to load so we don't get act()
-    // warnings.  The only thing rendered by App is children components.
-    // So await for those to be done with side effects (hooks).
-    // TODO: Render with new router
-    // await screen.findByText('Suite One');
+    console.log('test');
 
+    expect(getTestSuites).toBeCalledTimes(1);
     // expect(postTestSessions).toBeCalledTimes(1);
   });
 });
