+++ package.json
@@ -10,6 +10,7 @@
     "@mui/lab": "^5.0.0-alpha.58",
     "@mui/material": "^5.2.2",
     "@mui/styles": "^5.2.2",
+    "history": "^5.3.0",
     "js-yaml": "^4.1.0",
     "notistack": "^2.0.8",
     "react": "^17.0.2",
